pyproject.toml
[tool.poetry]
name = "chronicle"
version = "0.1.0"
description = ""
authors = ["zi-bot <andis95.muslim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
Faker = "^13.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
